+++ addons/yat/src/commands/View.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Collections.Generic;
+using Godot;
 using YAT.Attributes;
 using YAT.Interfaces;
-using YAT.Scenes;
 using YAT.Types;
 using static Godot.RenderingServer;
 
@@ -10,47 +10,41 @@ namespace YAT.Commands;
 
 [Command(
 	"view",
-	"Changes the rendering mode of the viewport.",
-	"[b]Usage[/b]: view [i]type[/i]\n\n" +
-	"Types:\n- normal: Disables debug draw.\n" +
-	"- unshaded: Renders the viewport without shading.\n" +
-	"- lightning: Enables lighting debug draw.\n" +
-	"- overdraw: Enables overdraw debug draw.\n" +
-	"- wireframe: Renders the viewport in wireframe mode.\n\n" +
-	"You can also use the integer value of the type " +
-	"[url=https://docs.godotengine.org/en/stable/classes/class_renderingserver.html#class-renderingserver-method-set-debug-generate-wireframes]ViewportDebugDraw[/url]."
+	"Changes the rendering mode of the viewport.\n"
+	+ "You can use either the name of the mode or the integer value.\n"
+	+ "See: [url=https://docs.godotengine.org/en/stable/classes/class_renderingserver.html#class-renderingserver-method-set-debug-generate-wireframes]ViewportDebugDraw[/url].",
+	"[b]Usage[/b]: view [i]type[/i]\n\n"
 )]
-[Argument("type", "normal|unshaded|lightning|overdraw|wireframe|int", "The type of debug draw to use.")]
+[Argument("mode", "string|int(0:)", "The type of debug draw to use.")]
 public sealed class View : ICommand
 {
-	private static readonly int MAX_DRAW_MODE = Enum.GetValues(typeof(ViewportDebugDraw)).Length - 1;
+	private static readonly int MAX_DRAW_MODE;
+	private static readonly List<StringName> Modes = new();
 
-	private static readonly Dictionary<string, ViewportDebugDraw> modeMappings = new()
-		{
-			{"normal", ViewportDebugDraw.Disabled},
-			{"unshaded", ViewportDebugDraw.Unshaded},
-			{"lightning", ViewportDebugDraw.Lighting},
-			{"overdraw", ViewportDebugDraw.Overdraw},
-			{"wireframe", ViewportDebugDraw.Wireframe}
-		};
+	static View()
+	{
+		var values = Enum.GetValues(typeof(ViewportDebugDraw));
+		MAX_DRAW_MODE = values.Length - 1;
+
+		foreach (var mode in values) Modes.Add(mode.ToString().ToLower());
+	}
 
 	private YAT _yat;
-	private BaseTerminal _terminal;
 
 	public CommandResult Execute(CommandData data)
 	{
-		var mode = data.RawData[1];
+		var mode = (string)data.Arguments["mode"];
 
 		_yat = data.Yat;
-		_terminal = data.Terminal;
 
-		if (modeMappings.TryGetValue(mode, out ViewportDebugDraw debugDraw))
-			return SetDebugDraw(debugDraw);
+		if (Modes.Contains(mode)) return SetDebugDraw(
+			(ViewportDebugDraw)Enum.Parse(typeof(ViewportDebugDraw), mode, true)
+		);
 
 		if (!int.TryParse(mode, out var iMode))
 			return ICommand.InvalidArguments($"Invalid mode: {mode}.");
 
-		if (!IsValidMode((ushort)iMode))
+		if (!IsValidMode(iMode))
 			return ICommand.InvalidArguments($"Invalid mode: {mode}. Valid range: 0 to {MAX_DRAW_MODE}.");
 
 		return SetDebugDraw((ViewportDebugDraw)iMode);
@@ -60,10 +54,8 @@ public sealed class View : ICommand
 	{
 		ViewportSetDebugDraw(_yat.GetViewport().GetViewportRid(), debugDraw);
 
-		_terminal.Print($"Set viewport debug draw to {debugDraw} ({(ushort)debugDraw}).");
-
-		return ICommand.Success();
+		return ICommand.Success($"Set viewport debug draw to {debugDraw} ({(int)debugDraw}).");
 	}
 
-	private bool IsValidMode(ushort mode) => mode >= 0 && mode <= MAX_DRAW_MODE;
+	private static bool IsValidMode(int mode) => mode >= 0 && mode <= MAX_DRAW_MODE;
 }